requirements.txt
PyGithub==2.1.1
python-dotenv==1.0.1
rich==13.7.0
matplotlib==3.8.2
numpy==1.26.3
textblob==0.17.1
requests==2.31.0 